requirements.txt
Flask~=1.1.2
python-dotenv==0.13.0
boto3==1.13.21
